[FE] rename: Modify folder names
