+++ frontend/src/components/community/friend-accept-button.tsx
