+++ frontend/src/components/community/friend-reject-button.tsx
