+++ frontend/src/components/community/friend-request-button.tsx
