+++ frontend/src/components/community/friend-request-item-skeleton.tsx
