+++ frontend/src/components/community/friend-request-list.tsx
@@ -1,7 +1,7 @@
 import type { User } from "@/types/user/type";
 import { cn } from "@/utils/style";
-import FriendRequestItem from "@/components/user/friend-request-item";
-import FriendRequestItemSkeleton from "@/components/user/friend-request-item-skeleton";
+import FriendRequestItem from "@/components/community/friend-request-item";
+import FriendRequestItemSkeleton from "@/components/community/friend-request-item-skeleton";
 
 interface FriendRequestListProps {
   isLoading: boolean;
