+++ frontend/src/components/community/profile-header.tsx
@@ -1,7 +1,7 @@
 import type { Profile } from "@/types/user/type";
 import defaultProfileBackground from "@/assets/images/profile-background.png";
-import UserPortrait from "@/components/user/user-portrait";
-import FriendRequestButton from "@/components/user/friend-request-button";
+import UserPortrait from "@/components/community/user-portrait";
+import FriendRequestButton from "@/components/community/friend-request-button";
 
 interface ProfileHeaderProps {
   profile: Profile;
