+++ frontend/src/components/community/user-item-skeleton.tsx
