+++ frontend/src/components/community/user-item.tsx
@@ -1,6 +1,6 @@
 import type { MouseEvent, ReactNode } from "react";
 
-import UserPortrait from "@/components/user/user-portrait";
+import UserPortrait from "@/components/community/user-portrait";
 import { cn } from "@/utils/style";
 import type { User } from "@/types/user/type";
