+++ frontend/src/components/community/user-search-modal.tsx
@@ -5,7 +5,7 @@ import AutocompleteSearchBar from "@/components/common/auto-complete-search-bar"
 import { getUserSearch } from "@/services/community/api";
 import type { User } from "@/types/user/type";
 import type { FriendSearchResponse } from "@/types/community/type";
-import UserSearchItem from "@/components/user/user-search-item";
+import UserSearchItem from "@/components/community/user-search-item";
 
 interface UserSearchModalProps extends DialogProps {
   keyword: string;
