+++ frontend/src/components/side-bar/side-bar-content.tsx
@@ -1,13 +1,13 @@
 import { useNavigate } from "react-router-dom";
 
-import UserItem from "@/components/user/user-item";
-import SettingButton from "@/components/user/setting-button";
+import UserItem from "@/components/community/user-item";
+import SettingButton from "@/components/community/setting-button";
 import SideBarFriendContent from "@/components/side-bar/side-bar-friend-content";
 import SideBarChatContent from "@/components/side-bar/side-bar-chat-content";
 import LogoutButton from "@/components/auth/logout-button";
 import type { SideBarTabType } from "@/types/side-bar/type";
 import useCurrentUser from "@/hooks/auth/use-current-user";
-import UserItemSkeleton from "@/components/user/user-item-skeleton";
+import UserItemSkeleton from "@/components/community/user-item-skeleton";
 
 interface SideBarContentProps {
   currentTab: SideBarTabType;
