--- frontend/src/types/user/type.ts
@@ -1,17 +1,0 @@
-export interface User {
-  userId: number;
-  nickname: string;
-  profileImageUrl: string | null;
-  online?: boolean;
-}
-
-export type FriendRequestType = "ACCEPTED" | "PENDING" | "ME" | "NONE";
-
-export interface Profile {
-  nickname: string;
-  profileImageUrl: string;
-  statusMessage: string;
-  friendCount: number;
-  joinedAt: string;
-  isFriend: FriendRequestType;
-}
